Initial checkin of ThermoAlert "Learning Project"
At this step, the application only has a “login” button that will log
to the Particle Cloud account.  The username and password are currently
hardcoded in the code under “loginAndDisplaysDevices” function, so to
make this work, don’t forget to put your registered account email and
password (careful to not push on git with that personal information if
you have a public account).